refactored threejs into its own component
